Update patch set 1
Patch Set 1:

Could you break this into two CLs, one that introduces the function and calls it for the existing semantics, the next which makes the second use of it?  This code is so tricky, I want to be very conservative.

Also, I am on vacation for another week.
Καλημερα σας.

Patch-set: 1
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>