Update patch set 1
Patch Set 1: Code-Review-1

I believe we didn't add these examples because they are in fact platform-dependent: An int may be 32 or 64bits, and the results depend on that size. So, running these examples (or seeing the results after running them) say on a 64bit platform may be confusing if the platform one is working on is a 32bit platform.

Also, these examples don't really add much to the understanding of the API. These are trivial functions.

Patch-set: 1
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1