Update patch set 1
Patch Set 1:

> Did you try the opposite approach? For each arch, use regexp.FindAllString to collect all `ssa.OpArch*`, create a lookup map, and then see if they're used. It sounds like it should be much faster, possibly enough that you won't need to parallelize it anymore.

I did try to make the process a bit smarter, but all in all, the files are big and regexp is pretty CPU-intensive. It helped a bit, but not as much as parallelizing the work.

I think it's still worth to do something along the lines of what you say, but we can tackle that later if rulegen is slow enough to bother anyone :)

Patch-set: 1
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>