Update patch set 2
Patch Set 2:

> I did try to make the process a bit smarter, but all in all, the files are big and regexp is pretty CPU-intensive. It helped a bit, but not as much as parallelizing the work.
> 
> I think it's still worth to do something along the lines of what you say, but we can tackle that later if rulegen is slow enough to bother anyone :)

Ugh, I was wrong; when I gave that a quick try yesterday, my code was buggy. Your suspicion is of course right; we have hundreds of ops per arch, so of course it's a huge speed-up. There's no noticeable gain by parallelizing the now tiny amount of CPU work.

Apologies for that mistake of mine. CL incoming.

Patch-set: 2