cmd/compile: add SSA rules for s390x compare-and-branch instructions
DO NOT SUBMIT, work in progress.

This commit adds SSA rules for the s390x combined compare-and-branch
instructions. These have a shorter encoding than separate compare
and branch instructions and they also don't clobber the condition
code (a.k.a. flag register).

I have deleted the 'loop_test.go' file and replaced it with a new
codegen test which performs a wider range of checks.

Change-Id: Ibca55d9c0aa1254aee69433731ab5d26a43a7c18