[dev.link] cmd/link: support new object file format
DO NOT REVIEW

Add support of parsing new object file format in the linker. At
least we can link a hello-world program.

The code here is probably not going in the final version. This
is basically only for debugging purposes -- to make sure the
writer and the reader work as expected.

Change-Id: I895aeea68326fabdb7e5aa1371b8cac7211a09dd