cmd/test2json: attribute output to the correct test
Fixes #29755

When printing regular test output check the indentation of the output, and use
the report stack to find the appropriate test name for that output.

This change includes a whitespace change to some golden test files. I suspect
this change is undesirable, however I am not sure how to proceed. From
what I can tell, at least with go1.13, the go binary outputs this indentation using
spaces not tabs. I'm not sure if that behaviour changed since go1.10 or if the
golden files are wrong.

The tabs in the golden files cause problems because the indentation check
looks for 4 spaces. If earlier versions of `go test` used tabs then I can update
the code to consider both tabs and space, and revert the changes to the
golden files. Please advise.

Change-Id: Iebab51816a9755168083a7a665b41497e9dfd85f
GitHub-Last-Rev: 271d3982783125bd34c67eaa029e72fdb128a354
GitHub-Pull-Request: golang/go#34419