net: add text marshalling and unmarshalling for HardwareAddr
The HardwareAddr type has the ability to be transformed to and from a
string. However, this capability is not exposed in a manner suitable
for use with various forms of marshalling and unmarshaling of text
(e.g. JSON or YAML). Let's add the proper functions so that
HardwareAddr implements the TextMarshaller and TextUnmarshaler
interfaces from the encoding package.

Fixes #29678

Change-Id: Id925ed9db08fc90788cb03ab3f77c9654aea262c
GitHub-Last-Rev: 85a3e003a6578c6c24c64e96128753d8e64be70d
GitHub-Pull-Request: golang/go#34452