cmd/compile: minor simplifications in rulegen
First, get rid of redundant parentheses in the generated code. This
trims only about 2300 bytes, but the code is a bit nicer to read.

Second, be consistent about declaring typ as &b.Func.Config.Types and
not &config.Types. Not particularly better, and it barely changes the
output, but we're more consistent now.

Third, remove a bit of duplication when handling the typ, auxint, and
aux variables.

Fourth and last, remove a stray canFail assignment; we ended up setting
that in add, not breakf, so it's not necessary to set it manually if we
don't use breakf.

Updates #33644.

Change-Id: I75999cb223a201969266fbfeae043599fa27fac5