encoding/xml: don't initialize struct fields when encoding
fieldInfo.value used to initialize nil anonymous struct fields if they
were encountered. This behavior is wanted when decoding, but not when
encoding. When encoding, the value should never be modified, and these
nil fields should be skipped entirely.

To fix the bug, add an "initialize" bool argument to the function, which
is false when encoding. In that case, we have nothing to encode, so
return reflect.Value{} to let the encoder skip the field.

Finally, add a couple of tests to cover the edge cases pointed out in
the original issue.

Fixes #27240.

Change-Id: Ic97ae4bfe5f2062c8518e03d1dec07c3875e18f6