Update patch set 1
Patch Set 1:

My guess is that when racecallbackthunk is called the value of R3 must always be 0, so it executes the few lines at the top of the function and returns. So the bad code would never get executed.

Patch-set: 1
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>