cmd/compile: don't emit autom's into object file
Don't write Autom records for when writing a function to the object
file; we no longer need functoin Autom's in the linker for DWARF
processing. So as to keep the object file format unchanged, write out
a zero-length list of automs to the object, as opposed to removing all
references.

Updates #34554.

Change-Id: I42a1d67207ea7114ae4f3a315cf37effba57f190