net: add text marshalling and unmarshalling for HardwareAddr
The HardwareAddr type has the ability to be transformed to and from a
string. However, this capability is not exposed in a manner suitable
for use with various forms of marshalling and unmarshaling of text
(e.g. JSON or YAML). Let's add the proper functions so that
HardwareAddr implements the TextMarshaller and TextUnmarshaler
interfaces from the encoding package.

Fixes #29678

Change-Id: Id925ed9db08fc90788cb03ab3f77c9654aea262c
GitHub-Last-Rev: df49700d7134cef359240e83f7c7c25cc9d7d58e
GitHub-Pull-Request: golang/go#34452