Update patch set 15
Patch Set 15:

(1 comment)

External linking on darwin/386 and linux/ppc64 are unsupported and need to be disabled on the link_syso_issue33139 test.

The linux/ppc* failure looks to be straightforward; I'll set up a qemu environment to test.

I think the resolution for AIX is the same as ppc on Linux; but there's no way I can test independently because I don't have access to AIX (commercial).  Is there any way I can get access to a buildbot to test it?

With the revert, I assume that I need to submit a new Change (CL) to Gerrit?

Patch-set: 15
Reviewer: Gerrit User 32979 <32979@62eb7196-b449-3ce5-99f1-c037f21e1705>