Update patch set 1
Patch Set 1:

> I'm not very familiar with the compiler's ssa so I'm not sure. What about all those b.Aux = nil lines?

One thing I was wondering is whether to have a method like:

// Reset removes the control values and any auxiliary values from the block
// and sets the blocks kind to the given value. The block's  successors will
// not be modified.
func (b *Block) Reset(kind BlockKind) {
    b.Kind
    b.ResetControls()
    b.Aux = nil
}

Which would also be useful if/when I add an AuxInt to Blocks since it could also set that to 0 (see CL 198037).

Ideally we'd also reset the successors and re-add them but that is very painful (OpPhis are aware of predecessor blocks so they need to be updated).

Patch-set: 1
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>