Add a cancel function which can cancel a context due to an error
Sometimes, we run mutilple go routines in a context, if one go routine
failed due to an error, we want to cancel the context and record the
error in it. So I added a function called WithErrorCancel function which
returns both CancelFunc and ErrorCancelFunc. CancelFunc is the same as
the function which is returned by WithCancel, ErrorCancelFunc is used to
cancel a context with an error.

Change-Id: Iebe058eb26f374247515f5ba67183c7fd068a6ef
GitHub-Last-Rev: b19f271b872672006df9b4520b2afe38ce861951
GitHub-Pull-Request: golang/go#34761