encoding/json: fix unmarshal to a map with key is user defined type
When unmarshal to a map, the map's key type must either be a string,
an integer, or implement encoding.TextUnmarshaler. But for a user
defined type, reflect.Kind can not distinguish between the static type
and the underlying type. In:

	var x MyString = "x"
	t := reflect.TypeOf(x)

the Kind of x is still reflect.String, even though the static type of x
is MyString.

Moreover, decodeState,object checks the map's key is string type has higher
priority than whether it implements encoding.TextUnmarshaler. So if
MyString implements encoding.TextUnmarshaler, its method is ignored when
decoding to map.

To fix this, just make checking implementing encoding.TextUnmarshaler
has the highest priority.

Fixes #34437

Change-Id: I780e0b084575e1dddfbb433fe03857adf71d05fb