Update patch set 3
Patch Set 3:

> The original issue had reservations from Joe, because some checks couldn't be done through UnmarshalJSON. Does your solution not have that problem? Please add a test case and mention it in the commit message.

Correct, this does not have that problem. Because the limit is not configurable or exposed via an API, there are no options to propagate to custom UnmarshalJSON implementers.

Because the maximum nesting depth enforcement is done by the scanner when it is ensuring the document is well-formed, prior to decoding and calling any UnmarshalJSON implementations, the protection is effective even in the presence of custom UnmarshalJSON implementations. Added a test case demonstrating this.

> Also, please briefly explain in the commit message how you arrived at the limit in this patch.

Done

> I think it would be better to just do two loops and one level of sub-tests

Done

Patch-set: 3